fix(stella-pipeline): give the pipeline path the cite_memory affordance it never had
`cite_memory` is registered on every surface and had never once been called:
zero `memory_citations` rows against 85 executions, with 8 recallable memories
present. That was read as a model-behaviour problem — a voluntary tool with no
in-turn moment that prompts it. It is not.

There are two renders of the same recalled frames, on two paths:

- `stella-cli`'s `render_context_section` (the REPL, `/goal`, the non-pipeline
  one-shot) emits `- [nod_…] label — content` AND asks for the citation. It has
  since it shipped.
- `assemble_user_message` — the pipeline, the DEFAULT `stella run` path —
  rendered `- [label] (source)` and dropped `RecalledFrame::id` on the floor,
  and asked for nothing.

So on the path those 85 executions came from the model was not declining to
cite. It could not: no `nod_…` id ever entered its context, `cite_memory`
rejects anything that is not `nod_` + 24 hex, and nothing asked. Selection
health is the fold over citations, so it stayed permanently empty — and
`retire_failing_context` can never fire without it. The context loop could
learn and never unlearn.

A materialized memory frame now carries its id beside the human label (L-C4:
the label still leads), and the block closes with the citation ask — once, and
only when some frame above is actually citable. Nothing citable, nothing asked:
inviting a citation with no handle to give would spend tokens soliciting a
fabricated id the tool would reject.

The ask is now one shared `ports::CITE_MEMORY_REQUEST` that both renders read.
Two copies of a prompt sentence is how one surface came to have none at all
while the other's existence read as evidence the affordance was wired.

Invariant 7 is untouched: recalled context on this path already rides as a
volatile message after the byte-stable system prefix, so no cached prefix
moves. The ask is not a demand — a turn is never failed for declining to cite,
which is the rule everywhere else in this plane.

Witness, checked the artisanal way (reverted the render, ran the test):
`a_recalled_memory_reaches_the_worker_citable` fails on the old render (no
`nod_…` in the assembled message) and passes here, with
`nothing_citable_means_nothing_is_asked` pinning the other side.
`recall_section_carries_ids_and_asks_for_citations` now asserts the CLI
surface renders the shared constant verbatim, so the two cannot re-diverge.

Three files were split to stay under the size ratchet, all mechanical:
`assemble_user_message` and its tests out of `pipeline.rs` (a god file closed
to growth) into `pipeline/user_message.rs`; `replay.rs`'s calibration tests
into `replay/calibration_tests.rs`; `wire_contract.rs`'s sample corpus into
`wire_contract/samples.rs`. No baseline entry was added or raised. The wire
sample now exercises the three #2194 channels with non-trivial values rather
than sampling them as zero.

`Refs`, not `Closes`: #2195's definition of done also asks for an end-to-end
witness that a citing turn produces a `memory_citations` row and a non-empty
`selection_health`. That spans four crates and is tracked separately.

Refs #2195
