+++ crates/stella-pipeline/src/pipeline.rs
@@ -127,6 +127,8 @@ mod scope_stage;
 mod stage_budget;
 mod task_frame;
 mod triage_stage;
+mod user_message;
+use user_message::{VerificationContract, assemble_user_message};
 mod verifier_stage;
 mod verify_probes;
 use verify_probes::DiffProbe;
@@ -3045,83 +3047,6 @@ fn bound_recalled_frames(frames: Vec<RecalledFrame>) -> Vec<RecalledFrame> {
         out.push(frame);
     }
     out
-}
-
-/// What the worker's user message says about how this run will be verified.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum VerificationContract<'a> {
-    /// An operator-configured oracle: disclose the command.
-    Oracle(&'a str),
-    /// No oracle and no independent witness author: the worker's own failing
-    /// test, written first, is the only deterministic evidence the run will
-    /// carry — say so on the channel the worker plans from.
-    WorkerTestFirst,
-    /// Nothing to add: a conversational turn, a class that never verifies, or
-    /// an authored witness that will supply the oracle post-execution (its
-    /// disclosure stays governed by the airlock).
-    None,
-}
-
-fn assemble_user_message(
-    goal: &str,
-    frames: &[RecalledFrame],
-    contract: VerificationContract<'_>,
-) -> String {
-    if frames.is_empty() && contract == VerificationContract::None {
-        return goal.to_string();
-    }
-    let mut s = String::new();
-    if !frames.is_empty() {
-        s.push_str("## Recalled context\n");
-        for f in frames {
-            // Cite by human label (L-C4); include content as grounding.
-            s.push_str("- [");
-            s.push_str(&f.citation_label);
-            s.push_str("] (");
-            s.push_str(&f.source);
-            s.push_str(")\n");
-            if !f.content.trim().is_empty() {
-                s.push_str("  ");
-                s.push_str(f.content.trim());
-                s.push('\n');
-            }
-        }
-        s.push('\n');
-    }
-    s.push_str("## Task\n");
-    s.push_str(goal.trim());
-    // The verification contract, when the operator configured one. The
-    // methodology prompt tells the worker to "run the target test" without
-    // ever saying which — the command that actually gates the run was
-    // withheld until the first failure disclosed it (the airlock's L1 brief
-    // names it anyway). Saying it up front moves that information one failed
-    // revision earlier, on the exact channel the worker plans from. Only the
-    // operator-CONFIGURED command is ever disclosed here: an authored
-    // witness's command does not exist yet at assembly time, and its
-    // disclosure stays governed by the airlock (`crate::witness::airlock`).
-    match contract {
-        VerificationContract::Oracle(command) => {
-            s.push_str("\n\n## Verification\n");
-            s.push_str(&format!(
-                "This run's primary verification is `{command}`: the accepted deterministic \
-                 evidence is this command failing before your change and passing after it. \
-                 Reproduce the failure with it before editing; make it pass before finishing. \
-                 Do not modify the tests it runs."
-            ));
-        }
-        VerificationContract::WorkerTestFirst => {
-            s.push_str("\n\n## Verification\n");
-            s.push_str(
-                "No test command is configured for this run and no independent test author \
-                 is available: nothing outside your own work will check this change. Before \
-                 implementing, write the failing test that captures this task and run it to \
-                 watch it fail; make it pass before finishing. That test is the only \
-                 deterministic evidence this run will carry.",
-            );
-        }
-        VerificationContract::None => {}
-    }
-    s
 }
 
 /// One digest over everything a model verdict depends on — goal, the (witness
