+++ crates/stella-pipeline/src/pipeline/tests.rs
@@ -1442,73 +1442,6 @@ async fn gather_diff_counts_real_new_file_lines_and_excludes_pre_existing() {
     assert!(probe3.text.contains("src/huge.rs"));
 }
 
-#[test]
-fn assemble_user_message_puts_recall_before_the_task() {
-    let frames = vec![RecalledFrame {
-        citation_label: "driver.rs".into(),
-        provider: "code-graph".into(),
-        source: "code-graph".into(),
-        kind: "symbol".into(),
-        uri: None,
-        method: None,
-        content: "run_turn".into(),
-        token_cost: 5,
-        id: None,
-        content_digest: None,
-    }];
-    let msg = assemble_user_message("do the thing", &frames, VerificationContract::None);
-    let recall_idx = msg.find("Recalled context").unwrap();
-    let task_idx = msg.find("do the thing").unwrap();
-    assert!(recall_idx < task_idx, "recall rides before the goal");
-}
-
-#[test]
-fn assemble_user_message_is_just_the_goal_when_no_recall() {
-    assert_eq!(
-        assemble_user_message("hello", &[], VerificationContract::None),
-        "hello"
-    );
-}
-
-/// The configured test command is the run's actual oracle, so the worker is
-/// told it up front instead of discovering it from the first failure's
-/// disclosure. Only the operator-configured command ever rides here — an
-/// authored witness's command is airlocked, and does not exist at assembly
-/// time anyway.
-#[test]
-fn assemble_user_message_states_the_configured_verification_contract() {
-    let msg = assemble_user_message(
-        "fix the parser",
-        &[],
-        VerificationContract::Oracle("cargo test -p parser"),
-    );
-    let task_idx = msg.find("fix the parser").unwrap();
-    let contract_idx = msg.find("`cargo test -p parser`").unwrap();
-    assert!(
-        task_idx < contract_idx,
-        "the task leads; the contract qualifies it"
-    );
-    assert!(msg.contains("failing before your change and passing after it"));
-    assert!(msg.contains("Do not modify the tests it runs"));
-}
-
-/// With no oracle and no independent author, the worker is told up front that
-/// its own failing test — written before the fix — is the run's only
-/// deterministic evidence. This is the same-model degraded posture: the run
-/// proceeds, but test-first is now the worker's job and the message says so.
-#[test]
-fn assemble_user_message_demands_test_first_when_nothing_else_verifies() {
-    let msg = assemble_user_message("add retries", &[], VerificationContract::WorkerTestFirst);
-    let task_idx = msg.find("add retries").unwrap();
-    let contract_idx = msg.find("write the failing test").unwrap();
-    assert!(
-        task_idx < contract_idx,
-        "the task leads; the contract qualifies it"
-    );
-    assert!(msg.contains("Before implementing"));
-    assert!(msg.contains("only deterministic evidence"));
-}
-
 /// With no `--test-command`, the witness author arms the flip oracle: its
 /// authored command is observed failing, the worker's change flips it, and
 /// the run submits fast on deterministic evidence — verifier skipped.
